pyproject.toml
[project]
name = "finbuzz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.6",
    "yfinance>=1.1.0",
]
